datapipeline/requirements.txt
pyspark==3.4.0
delta-spark==2.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
logging 